+++ .rspec
@@ -1,0 +1,1 @@
+--color
