+++ RFCEmailValidator.gemspec
@@ -6,7 +6,10 @@ Gem::Specification.new do |s|
   s.description = "Email validator based on the email RFC"
   s.authors     = ["Marton Meszaros"]
   s.email       = 'meza@meza.hu'
-  s.files       = ["lib/rfc_email_validator.rb"]
-  s.homepage    =
-      'http://rubygems.org/gems/RFCEmailValidator'
+  s.files       = `git ls-files`.split("\n").reject {|path| path =~ /\.gitignore$/ || path =~ /file$/ }
+  s.test_files  = `git ls-files -- {test,spec,features}/*`.split("\n")
+  s.homepage    = 'http://rubygems.org/gems/RFCEmailValidator'
+
+  s.add_development_dependency('rspec', '~> 2.12')
+  s.add_development_dependency('rake' , '>= 10.0.3')
 end
