+++ spec/rfc_email_validator/rfc_email_validator_spec.rb
@@ -1,0 +1,11 @@
+require 'spec_helper'
+
+describe RFCEmailValidator do
+	it "should return true for a 'name@example.org' like address" do
+		RFCEmailValidator.valid?("name@example.org").should be_true
+	end
+
+	it "should return false for a 'name@name@example.org' like address" do
+		RFCEmailValidator.valid?("name@name@example.org").should be_false
+	end
+end
