+++ spec/spec_helper.rb
@@ -1,0 +1,2 @@
+require 'rspec'
+require File.expand_path('../../lib/rfc_email_validator', __FILE__)
